backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "AI RAG API with FastAPI and LangChain"
authors = ["Andrea Ragalzi <andrea.ragalzi.code@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.121.1"
uvicorn = { extras = ["standard"], version = "^0.38.0" }
pydantic-settings = "^2.11.0"
langchain = "^1.0.5"
langchain-openai = "^1.0.2"
pypdf = "^6.2.0"
chromadb = "^1.3.4"
langchain-text-splitters = "^1.0.0"
langchain-community = "^0.4.1"
python-multipart = "^0.0.20"
cohere = "^5.20.0"
rank-bm25 = "^0.2.2"
unstructured = { extras = ["pdf"], version = "^0.18.18" }
langchain-chroma = "^1.0.0"
langdetect = "^1.0.9"
translate = "^3.8.0"
sendgrid = "^6.12.5"
python-dotenv = "^1.2.1"
firebase-admin = "^7.1.0"
aiofiles = "^25.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"
pytest-asyncio = "^1.3.0"
mypy = "^1.19.0"
pylint = "^3.3.9"
lizard = "^1.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Type checking - Firebase Admin SDK has incomplete stubs
[tool.pyright]
ignore = ["**/__pycache__", "**/chroma_db"]
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportGeneralTypeIssues = false

# MyPy configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
exclude = [
    "chroma_db",
    "__pycache__",
    ".venv",
]

# Ignore missing imports for third-party libraries without stubs
[[tool.mypy.overrides]]
module = [
    "langdetect",
    "translate",
    "sendgrid",
    "sendgrid.helpers.mail",
    "firebase_admin",
    "firebase_admin.*",
    "google.cloud.firestore",
]
ignore_missing_imports = true
